+++ salesforce/backend/__init__.py
@@ -40,10 +40,11 @@ DJANGO_40_PLUS = django.VERSION[:2] >= (4, 0)
 DJANGO_41_PLUS = django.VERSION[:2] >= (4, 1)
 DJANGO_42_PLUS = django.VERSION[:2] >= (4, 2)
 DJANGO_50_PLUS = django.VERSION[:2] >= (5, 0)
-max_django = (5, 0)
+DJANGO_51_PLUS = django.VERSION[:2] >= (5, 1)
+max_django = (5, 1)
 is_dev_version = django.VERSION[3:] and re.match('(alpha|beta|rc)', django.VERSION[3])
 if django.VERSION[:2] < (2, 1) or django.VERSION[:2] > max_django and not is_dev_version:
-    raise ImportError("Django version between 2.1 and 5.0 is required "
+    raise ImportError("Django version between 2.1 and 5.1 is required "
                       "for this django-salesforce.")
     # Usually three or more blocking issues can be expected by every
     # new major Django version. Strict check before support is better.